+++ client/src/hooks/useToast/useToast.tsx
@@ -1,13 +1,50 @@
-import {useContext} from 'react';
+/** @jsxImportSource @emotion/react */
+import {useEffect, useState} from 'react';
 
-import {ToastContext} from './ToastProvider';
+import {ToastMessage, ToastOptions, ToastArgs} from 'types/toastType';
+
+import toastEventManager from './toastEventManager';
+import {TOAST_EVENT} from './toastEventManager.type';
+
+const DEFAULT_TIME = 3000;
+
+type Toast = {
+  message: ToastMessage;
+  options: ToastOptions;
+};
 
 export const useToast = () => {
-  const context = useContext(ToastContext);
+  const [currentToast, setCurrentToast] = useState<ToastArgs | null>(null);
 
-  if (!context) {
-    throw new Error('useToast는 ToastProvider 내에서 사용되어야 합니다.');
-  }
+  const showToast = (message: ToastMessage, options: ToastOptions) => {
+    setCurrentToast({message, options});
+  };
 
-  return context;
+  const closeToast = () => {
+    setCurrentToast(null);
+  };
+
+  const setAutoCloseTimer = () => {
+    if (!currentToast) return;
+
+    const showingTime = currentToast.options.showingTime || DEFAULT_TIME;
+    const timer = setTimeout(() => setCurrentToast(null), showingTime);
+
+    return () => clearTimeout(timer);
+  };
+
+  useEffect(() => {
+    if (currentToast?.options.isAutoClosed) setAutoCloseTimer();
+
+    return;
+  }, [currentToast]);
+
+  useEffect(() => {
+    toastEventManager.addEventHandler({eventType: TOAST_EVENT.show, callback: showToast});
+  }, []);
+
+  return {
+    currentToast,
+    closeToast,
+  };
 };
